feat: remove resourceName from input
Because the most common use-case is to actually import everything. We
might end up needing to filter things later, but ultimately a static
string is too restrictive (eg, we can't import resources if their name
is dynamically generated when rendering).

When revisiting this, a more robust way to filter resources should be
implemented.

Signed-off-by: Lucas Caparelli <lucas.caparelli@gympass.com>